дз №3
