+++ Students/iessei-developer/project/webpack.config.js
@@ -12,7 +12,7 @@ module.exports = {
 
 		publicPath: 'dist',
 
-		filename: 'main.js'
+		filename: 'main1.js'
 
 	},
